create err res

+++ api/models/Problems.js
@@ -44,7 +44,7 @@ module.exports = {
 			defaultsTo:[]
 		},
 		time: {
-			type: "double",
+			type: "integer",
 			defaultsTo: Date.now()
 		}
